requirements.txt
scikit-surprise
numpy
pandas
matplotlib
seaborn
tensorflow